"Page modes" are now assigned on the fly
Using markdown metadata one can specify alternative modes, for instance:

    Mode: play
          edit

It will creates a radio button set allowing one to select the mode in
which to view the page, here with play mode as the default and edit mode
as the other option.

All it does is create the radio button set and assigning to the canvas
element (the space containing the annotations) a class equal to the
specified mode. In the case above this would be:

    <article id="canvas" class="ui-layout-center play">

By clicking the edit button it becomes:

    <article id="canvas" class="ui-layout-center edit">

The behaviour is up to the stylesheets and the javascript.