Resolved symbolic links in assets
Symbolic links are brittle and can break in versioning.
(which happened to me).

The original reasoning to have the link is to maintain some version
metadata contained in the filename of minified JavaScripts.
A more robust system would be to work with the unminified JavaScripts
in development, and use i.e. Django Compressor to minify and
concatenate the scripts for production.

The static/libs folder is quite messy containing a lot of extra stuff
from the original packages. Also most of the static assets between
aacore and aawiki are duplicated. Some spring cleaning ahead!